# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/workspace/users": {
      "get": {
        "operationId": "listWorkspaceUsers",
        "parameters": [
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "description": "Id from selected workspace",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Workspace ⌁ Users"
        ]
      },
      "post": {
        "operationId": "createWorkspaceUser",
        "parameters": [
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "description": "Id from selected workspace",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Workspace ⌁ Users"
        ]
      }
    },
    "/api/v1/workspace/users/{userId}": {
      "delete": {
        "operationId": "removeUserFromWorkspace",
        "parameters": [
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "description": "Id from selected workspace",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Workspace ⌁ Users"
        ]
      }
    },
    "/api/v1/workspace/users/invite": {
      "post": {
        "operationId": "inviteUserToWorkspace",
        "parameters": [
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "description": "Id from selected workspace",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteUserToWorkspaceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Workspace ⌁ Users"
        ]
      }
    },
    "/api/v1/workspace/users/invite/{inviteId}/join": {
      "post": {
        "operationId": "addUserToWorkspaceWithInvite",
        "parameters": [
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "description": "Id from selected workspace",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "inviteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Workspace ⌁ Users"
        ]
      }
    },
    "/api/v1/workspace/brands": {
      "get": {
        "description": "List all brands from a given workspace.",
        "operationId": "listBrands",
        "parameters": [
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "description": "Id from selected workspace",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageNumber",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "segment",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "SUPERMARKET",
                "EVENTS",
                "STORE",
                "PHARMACY",
                "CLINIC",
                "RESTAURANT",
                "REAL_ESTATE",
                "AUTOMOTIVE",
                "RADIO",
                "OTHER"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Brands listed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListBrandsResponseDto"
                }
              }
            }
          }
        },
        "summary": "List brands",
        "tags": [
          "Workspace ⌁ Brands"
        ]
      },
      "post": {
        "description": "Create a new brand.",
        "operationId": "createBrand",
        "parameters": [
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "description": "Id from selected workspace",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBrandDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Brand created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateBrandResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Brand name is taken.",
            "content": {
              "application/json": {
                "example": {
                  "message": "Brand name is taken.",
                  "error": "BRAND_NAME_IS_TAKEN",
                  "statusCode": 400
                }
              }
            }
          },
          "404": {
            "description": "Workspace not found.",
            "content": {
              "application/json": {
                "example": {
                  "message": "Workspace not found.",
                  "error": "WORKSPACE_NOT_FOUND",
                  "statusCode": 404
                }
              }
            }
          }
        },
        "summary": "Create brand",
        "tags": [
          "Workspace ⌁ Brands"
        ]
      }
    },
    "/api/v1/workspace/brands/{brandId}": {
      "put": {
        "description": "Update a brand (Do not include addresses infos).",
        "operationId": "updateBrand",
        "parameters": [
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "description": "Id from selected workspace",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "brandId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBrandDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Brand updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateBrandResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Brand name is taken.",
            "content": {
              "application/json": {
                "example": {
                  "message": "Brand name is taken.",
                  "error": "BRAND_NAME_IS_TAKEN",
                  "statusCode": 400
                }
              }
            }
          }
        },
        "summary": "Update brand",
        "tags": [
          "Workspace ⌁ Brands"
        ]
      }
    },
    "/api/v1/workspace/brands/{brandId}/address": {
      "post": {
        "description": "Add a new addresss to a brand.",
        "operationId": "createBrandAddress",
        "parameters": [
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "description": "Id from selected workspace",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "brandId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBrandAddressDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Address created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateBrandAddressResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Brand not found.",
            "content": {
              "application/json": {
                "example": {
                  "message": "Brand not found.",
                  "error": "BRAND_NOT_FOUND",
                  "statusCode": 404
                }
              }
            }
          }
        },
        "summary": "Create address",
        "tags": [
          "Workspace ⌁ Brands"
        ]
      },
      "delete": {
        "description": "Delete brand addresses for given ids.",
        "operationId": "deleteBrandAddresses",
        "parameters": [
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "description": "Id from selected workspace",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteBrandAddressesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Addresses deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteBrandAddressesResponseDto"
                }
              }
            }
          }
        },
        "summary": "Delete address",
        "tags": [
          "Workspace ⌁ Brands"
        ]
      }
    },
    "/api/v1/workspace/brands/{brandId}/address/{addressId}": {
      "put": {
        "description": "Add a new addresss to a brand.",
        "operationId": "updateBrandAddress",
        "parameters": [
          {
            "name": "X-Workspace-Id",
            "in": "header",
            "description": "Id from selected workspace",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "addressId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBrandAddressDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Address updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateBrandAddressResponseDto"
                }
              }
            }
          }
        },
        "summary": "Update address",
        "tags": [
          "Workspace ⌁ Brands"
        ]
      }
    },
...
  },
  "info": {
    "title": "WeSpot API",
    "description": "Endpoints for WeSpot API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "schemas": {
...
      "CreateWorkspaceUserDto": {
        "type": "object",
        "properties": {}
      },
      "InviteUserToWorkspaceDto": {
        "type": "object",
        "properties": {}
      },
      "PaginationResponseDto": {
        "type": "object",
        "properties": {
          "totalItems": {
            "type": "number"
          },
          "pageNumber": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          },
          "hasPrevPage": {
            "type": "boolean"
          },
          "hasNextPage": {
            "type": "boolean"
          }
        },
        "required": [
          "totalItems",
          "pageNumber",
          "pageSize",
          "totalPages",
          "hasPrevPage",
          "hasNextPage"
        ]
      },
      "ListBrandsResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BrandEntityDto"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationResponseDto"
          }
        },
        "required": [
          "items",
          "pagination"
        ]
      },
      "CreateBrandAddressResponsibleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "role": {
            "type": "object",
            "nullable": true
          }
        }
      },
      "CreateBrandAddressDto": {
        "type": "object",
        "properties": {
          "state": {
            "type": "string",
            "enum": [
              "AC",
              "AL",
              "AP",
              "AM",
              "BA",
              "CE",
              "DF",
              "ES",
              "GO",
              "MA",
              "MT",
              "MS",
              "MG",
              "PA",
              "PB",
              "PR",
              "PE",
              "PI",
              "RJ",
              "RN",
              "RS",
              "RO",
              "RR",
              "SC",
              "SP",
              "SE",
              "TO"
            ],
            "nullable": true
          },
          "city": {
            "type": "object",
            "nullable": true
          },
          "neighborhood": {
            "type": "object",
            "nullable": true
          },
          "street": {
            "type": "object",
            "nullable": true
          },
          "number": {
            "type": "object",
            "nullable": true
          },
          "complement": {
            "type": "object",
            "nullable": true
          },
          "responsibles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateBrandAddressResponsibleDto"
            }
          }
        },
        "required": [
          "responsibles"
        ]
      },
      "CreateBrandDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "segment": {
            "type": "string",
            "enum": [
              "SUPERMARKET",
              "EVENTS",
              "STORE",
              "PHARMACY",
              "CLINIC",
              "RESTAURANT",
              "REAL_ESTATE",
              "AUTOMOTIVE",
              "RADIO",
              "OTHER"
            ]
          },
          "description": {
            "type": "object",
            "nullable": true
          },
          "slogan": {
            "type": "object",
            "nullable": true
          },
          "phoneNumber": {
            "type": "object",
            "nullable": true
          },
          "whatsapp": {
            "type": "object",
            "nullable": true
          },
          "instagram": {
            "type": "object",
            "nullable": true
          },
          "tiktok": {
            "type": "object",
            "nullable": true
          },
          "facebook": {
            "type": "object",
            "nullable": true
          },
          "website": {
            "type": "object",
            "nullable": true
          },
          "addresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateBrandAddressDto"
            }
          }
        },
        "required": [
          "name",
          "segment",
          "addresses"
        ]
      },
      "CreateBrandResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "segment": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "logoKey": {
            "type": "string",
            "nullable": true
          },
          "slogan": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "whatsapp": {
            "type": "string",
            "nullable": true
          },
          "instagram": {
            "type": "string",
            "nullable": true
          },
          "tiktok": {
            "type": "string",
            "nullable": true
          },
          "facebook": {
            "type": "string",
            "nullable": true
          },
          "website": {
            "type": "string",
            "nullable": true
          },
          "spots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpotEntityDto"
            }
          },
          "addresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BrandAddressEntityDto"
            }
          },
          "workspaceId": {
            "type": "string"
          },
          "workspace": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "name",
          "segment",
          "description",
          "logoKey",
          "slogan",
          "phoneNumber",
          "whatsapp",
          "instagram",
          "tiktok",
          "facebook",
          "website",
          "spots",
          "addresses",
          "workspaceId",
          "workspace"
        ]
      },
      "UpdateBrandDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "segment": {
            "type": "string",
            "enum": [
              "SUPERMARKET",
              "EVENTS",
              "STORE",
              "PHARMACY",
              "CLINIC",
              "RESTAURANT",
              "REAL_ESTATE",
              "AUTOMOTIVE",
              "RADIO",
              "OTHER"
            ]
          },
          "description": {
            "type": "object",
            "nullable": true
          },
          "slogan": {
            "type": "object",
            "nullable": true
          },
          "phoneNumber": {
            "type": "object",
            "nullable": true
          },
          "whatsapp": {
            "type": "object",
            "nullable": true
          },
          "instagram": {
            "type": "object",
            "nullable": true
          },
          "tiktok": {
            "type": "object",
            "nullable": true
          },
          "facebook": {
            "type": "object",
            "nullable": true
          },
          "website": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "name",
          "segment"
        ]
      },
      "UpdateBrandResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "segment": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "logoKey": {
            "type": "string",
            "nullable": true
          },
          "slogan": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "whatsapp": {
            "type": "string",
            "nullable": true
          },
          "instagram": {
            "type": "string",
            "nullable": true
          },
          "tiktok": {
            "type": "string",
            "nullable": true
          },
          "facebook": {
            "type": "string",
            "nullable": true
          },
          "website": {
            "type": "string",
            "nullable": true
          },
          "spots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpotEntityDto"
            }
          },
          "addresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BrandAddressEntityDto"
            }
          },
          "workspaceId": {
            "type": "string"
          },
          "workspace": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "name",
          "segment",
          "description",
          "logoKey",
          "slogan",
          "phoneNumber",
          "whatsapp",
          "instagram",
          "tiktok",
          "facebook",
          "website",
          "spots",
          "addresses",
          "workspaceId",
          "workspace"
        ]
      },
      "CreateBrandAddressResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "state": {
            "type": "object"
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "neighborhood": {
            "type": "string",
            "nullable": true
          },
          "street": {
            "type": "string",
            "nullable": true
          },
          "number": {
            "type": "string",
            "nullable": true
          },
          "complement": {
            "type": "string",
            "nullable": true
          },
          "brandId": {
            "type": "string"
          },
          "brand": {
            "type": "object"
          },
          "responsibles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BrandResponsibleEntityDto"
            }
          }
        },
        "required": [
          "id",
          "state",
          "city",
          "neighborhood",
          "street",
          "number",
          "complement",
          "brandId",
          "brand",
          "responsibles"
        ]
      },
...
      "UpdateBrandAddressDto": {
        "type": "object",
        "properties": {
          "state": {
            "type": "string",
            "enum": [
              "AC",
              "AL",
              "AP",
              "AM",
              "BA",
              "CE",
              "DF",
              "ES",
              "GO",
              "MA",
              "MT",
              "MS",
              "MG",
              "PA",
              "PB",
              "PR",
              "PE",
              "PI",
              "RJ",
              "RN",
              "RS",
              "RO",
              "RR",
              "SC",
              "SP",
              "SE",
              "TO"
            ],
            "nullable": true
          },
          "city": {
            "type": "object",
            "nullable": true
          },
          "neighborhood": {
            "type": "object",
            "nullable": true
          },
          "street": {
            "type": "object",
            "nullable": true
          },
          "number": {
            "type": "object",
            "nullable": true
          },
          "complement": {
            "type": "object",
            "nullable": true
          },
          "responsibles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UpdateBrandAddressResponsibleDto"
            }
          }
        },
        "required": [
          "responsibles"
        ]
      },
...
      "DeleteBrandAddressesDto": {
        "type": "object",
        "properties": {
          "ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "ids"
        ]
      },
      "DeleteBrandAddressesResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ]
      },
...
    }
  }
}